refactor: migrate settings feature to feature slice
- Move SettingsContext and SettingsProvider to src/features/settings/
- Move useSettings hook to src/features/settings/hooks/
- Move all 14 settings components to src/features/settings/components/
- Create barrel exports (index.ts) for clean public API
- Update imports in Settings.tsx to use @/features/settings
- Create backward-compatible re-exports in old locations

All 879 tests pass.
