+++ src/features/settings/components/ClearDataButton.test.tsx
