+++ src/features/settings/components/ClearDataButton.tsx
