+++ src/features/settings/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useSettings } from './useSettings';
