+++ src/features/settings/hooks/useSettings.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { SettingsContext } from '../context';
+
+export function useSettings() {
+  const context = useContext(SettingsContext);
+  if (!context) {
+    throw new Error('useSettings must be used within SettingsProvider');
+  }
+  return context;
+}
