+++ src/features/settings/index.ts
@@ -1,0 +1,27 @@
+// Context
+export { SettingsContext } from './context';
+export type { SettingsContextValue } from './context';
+
+// Provider
+export { SettingsProvider } from './provider';
+
+// Hooks
+export { useSettings } from './hooks';
+
+// Components
+export {
+  ClearDataButton,
+  DebugExport,
+  DisabledRecommendations,
+  ExportButton,
+  ExportRecommendationsButton,
+  HouseholdForm,
+  ImportButton,
+  ImportRecommendationsButton,
+  LanguageSelector,
+  NutritionSettings,
+  OverriddenRecommendations,
+  RecommendationsStatus,
+  ShoppingListExport,
+  ThemeSelector,
+} from './components';
