+++ src/shared/hooks/useSettings.ts
@@ -1,10 +1,4 @@
-import { useContext } from 'react';
-import { SettingsContext } from '@/shared/contexts/SettingsContext';
-
-export function useSettings() {
-  const context = useContext(SettingsContext);
-  if (!context) {
-    throw new Error('useSettings must be used within SettingsProvider');
-  }
-  return context;
-}
+/**
+ * @deprecated Import from '@/features/settings' instead
+ */
+export { useSettings } from '@/features/settings';
